requirements.txt
gradio==3.47.1
opencv-contrib-python==4.8.0.74
dlib
piexif
gfpgan==1.3.8
safetensors==0.3.3
psutil==5.9.5
einops==0.6.1
transformers==4.26.1
dynamicprompts
scipy==1.11.2
torchsde==0.2.5
tqdm==4.66.1
numpy==1.23.5
scikit-image
onnxruntime
accelerate==0.22.0
insightface==0.7.3
facexlib==0.3.0